ariel_os_sensors

Enum Category

#[non_exhaustive]
pub enum Category {

    Accelerometer,
    AccelerometerTemperature,
    AccelerometerGyroscope,
    AccelerometerGyroscopeTemperature,
    AccelerometerMagnetometerTemperature,
    Ammeter,
    Co2Gas,
    Color,
    Gnss,
    Gyroscope,
    RelativeHumidity,
    RelativeHumidityTemperature,
    Light,
    Magnetometer,
    Ph,
    Pressure,
    PressureTemperature,
    PushButton,
    Temperature,
    Tvoc,
    Voltage,

}

Categories a sensor driver can be part of.

A sensor driver may be part of multiple categories.

For sensor driver implementors

Many MEMS sensor devices (e.g., accelerometers) include a temperature sensing element in addition to their main sensing element, as temperature may slightly affect the measurement results. Sensor drivers are not under the obligation of exposing such temperature readings, even if they are exposed by the sensor device. They may however still be fetched by the sensor driver internally, especially to dynamically compute the accuracy of the main reading returned by the sensor driver. If temperature readings are not exposed by the sensor driver, the sensor driver must not be considered part of a category that includes temperature (Category::Temperature or Category::AccelerometerTemperature in the case of an accelerometer), even if the sensor device does expose them. One reason to not expose the extra readings is a lack of precision or accuracy with the extra sensing element.

Sensor drivers may be part of multiple categories and should then list all of them: e.g., being part of the Category::AccelerometerTemperature does not imply also being part of the Category::Accelerometer category, and the sensor driver must list both of them.

Missing variants can be added when required. Please open an issue to discuss it.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Accelerometer

Accelerometer.

AccelerometerTemperature

Accelerometer & temperature sensor.

AccelerometerGyroscope

Accelerometer & gyroscope, also known as inertial measurement unit (IMU).

AccelerometerGyroscopeTemperature

Accelerometer & gyroscope & temperature sensor, also known as inertial measurement unit (IMU).

AccelerometerMagnetometerTemperature

Accelerometer & magnetometer & temperature sensor.

Ammeter

Ammeter (ampere meter).

Co2Gas

CO₂ gas sensor.

Color

Color sensor.

Gnss

GNSS (Global Navigation Satellite System) receiver.

Gyroscope

Gyroscope.

RelativeHumidity

Relative humidity sensor.

RelativeHumidityTemperature

Relative humidity & temperature sensor.

Light

Light sensor.

Magnetometer

Magnetometer.

Ph

pH sensor.

Pressure

Pressure sensor.

PressureTemperature

Pressure & temperature sensor.

PushButton

Push button.

Temperature

Temperature sensor.

Tvoc

TVOC sensor.

Voltage

Voltage sensor.